Retweets use original retweeted text.
Certain tweets are longer than 140 characters; notably, retweets that contain a link may result in long tweets that are truncated by Twitter API for compatibility reasons. The original retweet text is available in the API, and this update uses that text. It also prepends the tweet with "RT @UserNameHere" to match Twitter's style.